moved popup functionality to another file but kept logic the same
