+++ src/ui/home.rs
@@ -2,13 +2,13 @@ use ratatui::{
     layout::{Constraint, Direction, Layout, Rect},
     style::{Color, Style},
     text::{Line, Span, Text},
-    widgets::{Block, Borders, Clear, Paragraph, Wrap},
+    widgets::{Block, Borders, Paragraph, Wrap},
     Frame,
 };
 
-use crate::app::{
-    ActivePanel, App, CurrentScreen, DetailField, Groups, HeaderInputMode, ParameterInputMode,
-    RequestType,
+use crate::app::{ActivePanel, App, CurrentScreen, DetailField, Groups};
+use crate::ui::popups::{
+    add_request_popup, editing_popup, exiting_popup, render_header_popup, render_params_popup,
 };
 
 pub fn ui(frame: &mut Frame, app: &mut App) {
@@ -23,72 +23,6 @@ pub fn ui(frame: &mut Frame, app: &mut App) {
     if app.adding_params {
         render_params_popup(frame, app);
     }
-}
-
-fn render_params_popup(frame: &mut Frame<'_>, app: &mut App) {
-    // First render a Clear widget over the area where the popup will be
-    let area = centered_rect(60, 25, frame.area());
-    frame.render_widget(Clear, area);
-
-    // Create and render the popup block
-    let popup_block = Block::default()
-        .title("Add Parameter")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::DarkGray));
-
-    frame.render_widget(popup_block, area);
-
-    let inner_area = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(1)
-        .constraints([
-            Constraint::Length(3), // Params key
-            Constraint::Length(3), // Params value
-            Constraint::Length(2), // Instructions
-        ])
-        .split(area);
-
-    // Header key input
-    let key_block = Block::default()
-        .title("Parameter Key")
-        .borders(Borders::ALL)
-        .border_style(
-            if matches!(app.params_input_mode, ParameterInputMode::Key) {
-                Style::default().fg(Color::Yellow)
-            } else {
-                Style::default()
-            },
-        );
-
-    let key_input = Paragraph::new(app.params_key_input.as_str())
-        .block(key_block)
-        .style(Style::default().fg(Color::White));
-    frame.render_widget(key_input, inner_area[0]);
-
-    // Header value input
-    let value_block = Block::default()
-        .title("Parameter Value")
-        .borders(Borders::ALL)
-        .border_style(
-            if matches!(app.params_input_mode, ParameterInputMode::Value) {
-                Style::default().fg(Color::Yellow)
-            } else {
-                Style::default()
-            },
-        );
-
-    let value_input = Paragraph::new(app.params_value_input.as_str())
-        .block(value_block)
-        .style(Style::default().fg(Color::White));
-    frame.render_widget(value_input, inner_area[1]);
-
-    // Instructions
-    let instructions = Paragraph::new(match app.params_input_mode {
-        ParameterInputMode::Key => "Enter params key (Enter/Tab to move to value)",
-        ParameterInputMode::Value => "Enter params value (Enter to save)",
-    })
-    .style(Style::default().fg(Color::Gray));
-    frame.render_widget(instructions, inner_area[2]);
 }
 
 fn render_base_ui(frame: &mut Frame, app: &mut App) {
@@ -508,33 +442,7 @@ fn render_base_ui(frame: &mut Frame, app: &mut App) {
 
     // Editing popup
     if let Some(Groups::Name) = app.groups {
-        let popup_block = Block::default()
-            .title("Enter group name")
-            .borders(Borders::NONE)
-            .style(Style::default().bg(Color::DarkGray));
-
-        let area = centered_rect(60, 25, frame.area());
-        frame.render_widget(Clear, area);
-        frame.render_widget(popup_block, area);
-
-        let input_block = Block::default()
-            .borders(Borders::ALL)
-            .style(Style::default().bg(Color::DarkGray));
-
-        let input = Paragraph::new(app.key_input.as_str())
-            .block(input_block)
-            .style(Style::default().fg(Color::White));
-
-        let input_area = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([
-                Constraint::Length(1),
-                Constraint::Length(3),
-                Constraint::Min(0),
-            ])
-            .split(area)[1];
-
-        frame.render_widget(input, input_area);
+        editing_popup(frame, app);
     }
 
     // TODO - add delete group popup
@@ -542,76 +450,16 @@ fn render_base_ui(frame: &mut Frame, app: &mut App) {
 
     // Exit popup
     if app.current_screen == CurrentScreen::Exiting {
-        let popup_block = Block::default()
-            .title("Quit")
-            .borders(Borders::NONE)
-            .style(Style::default().bg(Color::DarkGray));
-
-        let area = centered_rect(60, 25, frame.area());
-        frame.render_widget(Clear, area);
-
-        let exit_text = Text::styled(
-            "Are you sure you want to quit? (y/n)",
-            Style::default().fg(Color::Red),
-        );
-        let exit_paragraph = Paragraph::new(exit_text)
-            .block(popup_block)
-            .wrap(Wrap { trim: false });
-
-        frame.render_widget(exit_paragraph, area);
+        exiting_popup(frame);
     }
 
     // Add Request popup
     if app.current_screen == CurrentScreen::AddingRequest {
-        let popup_block = Block::default()
-            .title("Add New Request")
-            .borders(Borders::ALL)
-            .style(Style::default().bg(Color::DarkGray));
-
-        let area = centered_rect(60, 25, frame.area());
-        frame.render_widget(Clear, area);
-        frame.render_widget(popup_block, area);
-
-        let inner_area = Layout::default()
-            .direction(Direction::Vertical)
-            .margin(1)
-            .constraints([
-                Constraint::Length(3), // Request name
-                Constraint::Length(3), // Request type
-                Constraint::Length(2), // Instructions
-            ])
-            .split(area);
-
-        // Request name input
-        let name_block = Block::default().title("Request Name").borders(Borders::ALL);
-        let name_input = Paragraph::new(app.request_name_input.as_str())
-            .block(name_block)
-            .style(Style::default().fg(Color::White));
-        frame.render_widget(name_input, inner_area[0]);
-
-        // Request type selection
-        let type_block = Block::default()
-            .title("Request Type (←/→ to change)")
-            .borders(Borders::ALL);
-        let type_text = Paragraph::new(app.selected_request_type.as_str())
-            .block(type_block)
-            .style(Style::default().fg(match app.selected_request_type {
-                RequestType::GET => Color::Green,
-                RequestType::POST => Color::Blue,
-                RequestType::PUT => Color::Yellow,
-                RequestType::DELETE => Color::Red,
-                RequestType::PATCH => Color::Magenta,
-            }));
-        frame.render_widget(type_text, inner_area[1]);
-
-        // Instructions
-        let instructions = Paragraph::new("Press Enter to save, Esc to cancel")
-            .style(Style::default().fg(Color::Gray));
-        frame.render_widget(instructions, inner_area[2]);
+        add_request_popup(frame, app);
     }
 }
 
-fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
+pub fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
     let popup_layout = Layout::default()
         .direction(Direction::Vertical)
         .constraints([
@@ -629,66 +477,4 @@ fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
             Constraint::Percentage((100 - percent_x) / 2),
         ])
         .split(popup_layout[1])[1]
-}
-
-fn render_header_popup(frame: &mut Frame, app: &App) {
-    // First render a Clear widget over the area where the popup will be
-    let area = centered_rect(60, 25, frame.area());
-    frame.render_widget(Clear, area);
-
-    // Create and render the popup block
-    let popup_block = Block::default()
-        .title("Add Header")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::DarkGray));
-
-    frame.render_widget(popup_block, area);
-
-    let inner_area = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(1)
-        .constraints([
-            Constraint::Length(3), // Header key
-            Constraint::Length(3), // Header value
-            Constraint::Length(2), // Instructions
-        ])
-        .split(area);
-
-    // Header key input
-    let key_block = Block::default()
-        .title("Header Key")
-        .borders(Borders::ALL)
-        .border_style(if matches!(app.header_input_mode, HeaderInputMode::Key) {
-            Style::default().fg(Color::Yellow)
-        } else {
-            Style::default()
-        });
-
-    let key_input = Paragraph::new(app.header_key_input.as_str())
-        .block(key_block)
-        .style(Style::default().fg(Color::White));
-    frame.render_widget(key_input, inner_area[0]);
-
-    // Header value input
-    let value_block = Block::default()
-        .title("Header Value")
-        .borders(Borders::ALL)
-        .border_style(if matches!(app.header_input_mode, HeaderInputMode::Value) {
-            Style::default().fg(Color::Yellow)
-        } else {
-            Style::default()
-        });
-
-    let value_input = Paragraph::new(app.header_value_input.as_str())
-        .block(value_block)
-        .style(Style::default().fg(Color::White));
-    frame.render_widget(value_input, inner_area[1]);
-
-    // Instructions
-    let instructions = Paragraph::new(match app.header_input_mode {
-        HeaderInputMode::Key => "Enter header key (Enter/Tab to move to value)",
-        HeaderInputMode::Value => "Enter header value (Enter to save)",
-    })
-    .style(Style::default().fg(Color::Gray));
-    frame.render_widget(instructions, inner_area[2]);
 }
