+++ src/ui/mod.rs
@@ -1,2 +1,7 @@
 mod home;
+mod popups;
+pub use home::centered_rect;
 pub use home::ui;
+pub use popups::{
+    add_request_popup, editing_popup, exiting_popup, render_header_popup, render_params_popup,
+};
